Now can pass the test1j
I changed the design of send/receive message to pass the test
Now no matter the message is sent immediately or not,
the message counter will increment. and the resumed process will
decrement the msg counter. Which can avoid the full queue error complained
in test1j, but in a meaningless way for an OS.